# Improve performance of collect using BangBang.collector