--- a/test/environments/main/Manifest.toml
+++ b/test/environments/main/Manifest.toml
@@ -21,7 +21,7 @@ version = "0.1.5"
 
 [[BangBang]]
 deps = ["Compat", "ConstructionBase", "Future", "InitialValues", "LinearAlgebra", "Requires", "Setfield", "Tables", "ZygoteRules"]
-git-tree-sha1 = "e9bff4d95421205dc7c1eb32cf8d403e7c7e2d52"
+git-tree-sha1 = "7e5f8c258ae1d8210d0329488e76a1816e68fb20"
 repo-rev = "master"
 repo-url = "https://github.com/tkf/BangBang.jl.git"
 uuid = "198e06fe-97b7-11e9-32a5-e1d131e6ad66"