https://github.com/tkf/BangBang.jl/pull/143

- [x] Require BangBang 0.3.21

## Commit Message
Improve performance of collect using BangBang.collector (#286)