@tkf 2020-05-21
PR e642869cb42535cb102adaddf3c37ba72c0670c5 with Julia 1.6.0-DEV.62:

```
3-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "identity-union" => Trial(337.158 μs)
  "identity-float" => Trial(17.091 μs)
  "filter-missing" => Trial(43.867 μs)
```

`master` 4b25dcc63d0cf91a73fa9538b95b1d10ea3330b0 with Julia 1.6.0-DEV.62:

```
3-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "identity-union" => Trial(349.913 μs)
  "identity-float" => Trial(54.675 μs)
  "filter-missing" => Trial(79.479 μs)
```

---

PR e642869cb42535cb102adaddf3c37ba72c0670c5 with Julia 1.4.1:

```
3-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "identity-union" => Trial(266.241 μs)
  "identity-float" => Trial(70.341 μs)
  "filter-missing" => Trial(92.778 μs)
```

`master` 4b25dcc63d0cf91a73fa9538b95b1d10ea3330b0 with Julia 1.4.1:

```
3-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "identity-union" => Trial(296.679 μs)
  "identity-float" => Trial(54.757 μs)
  "filter-missing" => Trial(77.499 μs)
```

---

So, like https://github.com/tkf/BangBang.jl/pull/143#issuecomment-631890246, this also needs Julia 1.6 (or maybe 1.5).