@tkf 2020-05-21
PR 232f8b80d1f05b9f34170653ef86ecf27f7c5614 with 1.6.0-DEV.62

```
3-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "identity-union" => Trial(327.300 μs)
  "identity-float" => Trial(17.288 μs)
  "filter-missing" => Trial(49.022 μs)
```

---

PR 232f8b80d1f05b9f34170653ef86ecf27f7c5614 with 1.4.1

```
3-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "identity-union" => Trial(256.708 μs)
  "identity-float" => Trial(68.152 μs)
  "filter-missing" => Trial(91.982 μs)
```